Move Output and add interface
